Correção no layout do dialog

+++ app/src/main/java/br/com/trivio/wms/components/custom/NumberInput.kt
@@ -66,7 +66,7 @@ class NumberInput @JvmOverloads constructor(
 
   fun setValue(value: BigDecimal?) {
     value?.let {
-      this.setValue(it.toString())
+      this.setValue(it.toInt().toString())
     }
   }
